Implement out-of-bound execution for OpenStack and Keystone clients
In order to better adhere to the OpenID Connect specs, the Keystone
server must be the OpenID Connect client (Relying Party). This means
that the auth flow should be handled by the Keystone app (the RP), and
not the OpenStack clients. The OpenID Connect requires that
the user authorizes the request in the OpenID Connect Provider (OP),
meaning that the RP must interact with the user-agent (the browser in
this case) for authentication.

However, in the case of CLI applications, it is not easy to perform this
authorization request. We implement out-of-bound authorization as
follows:

- The OpenStack CLI makes a post request to the Keystone Federation
  endpoint (RP), configured to use OpenID Connect, specifying a special
  parameter (oscli=1) in the query string. This indicates out-of-bound
  execution.

- The Keystone Federation endpoint redirects the user to the OP in order
  to authorize the request. The redirect_uri that is passed to the OP so
  that the auth request callback is performed is not a URL from the RP,
  but http://localhost:8080

- The OpenStack CLI spawns a http server in localhost:8080 and it waits
  for the authorization code.

- The user authorizes the request in the OP, then it is redirected to
  http://localhost:8080

- The OpenStack CLI gets the code, and completes the auth request, by
  contacting again the Keystone Federation endpoint with the auth code.